requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.2
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
